txnid2: introduce data skew into BigTableLoader

+++ tests/test_apps/txnid-selfcheck2/src/txnIdSelfCheck/LoadTableLoader.java
@@ -82,7 +82,7 @@ public class LoadTableLoader extends Thread {
     final BlockingQueue<Long> cpDelQueue = new LinkedBlockingQueue<Long>();
     final BlockingQueue<Long> onlyDelQueue = new LinkedBlockingQueue<Long>();
 
-    LoadTableLoader(Client client, String tableName, int targetCount, int batchSize, Semaphore permits, boolean isMp, int pcolIdx)
+    LoadTableLoader(Client client, String tableName, long targetCount, int batchSize, Semaphore permits, boolean isMp, int pcolIdx)
             throws IOException, ProcCallException {
         setName("LoadTableLoader-" + tableName);
         setDaemon(true);
